Вынес карту в кастомный хук
